+++ src/odoo_configurator/configurator.py
@@ -29,6 +29,7 @@ from .apps import import_configurator
 from .import_manager import ImportManager
 from .logging import get_logger
 from .odoo_connection import OdooConnection
+from .odoo_connection import get_dir_full_path
 from .keepass import KeepassCli
 from .bitwarden import Bitwarden
 
@@ -89,6 +90,7 @@ class Configurator:
             http_password=odoo_params.get('http_password'),
             createdb=odoo_params.get('create_db'),
             debug_xmlrpc=self.debug_xmlrpc,
+            configurator=self
         )
 
     def get_odoo_auth_params(self):
@@ -139,7 +141,7 @@ class Configurator:
             script_files = self.get_files_path(parsed_config['script_files'])
 
             for script_file in script_files:
-                parsed_script = get_config_from_files(script_files)
+                parsed_script = get_config_from_files([script_file])
                 if parsed_script.get('title'):
                     parsed_script['title'] = '%s : %s' % (os.path.basename(script_file),
                                                           parsed_script.get('title'))
@@ -151,13 +153,17 @@ class Configurator:
 
     def get_release_files(self):
         files = []
-        if os.path.isdir(self.release_directory):
-            files = glob.glob(os.path.join(self.release_directory, '*.yml'))
+        release_dir = self.release_directory
+        if not os.path.isdir(release_dir):
+            release_dir = os.path.join(os.path.dirname(sys.argv[1]), release_dir)
+        if os.path.isdir(release_dir):
+            files = glob.glob(os.path.join(release_dir, '*.yml'))
         return files
 
     def backup_release_directory(self):
-        if self.clear_release_directory and os.path.isdir(self.release_directory):
-            bak_dir = os.path.join(self.release_directory, 'bak')
+        release_directory = get_dir_full_path(self.release_directory)
+        if self.clear_release_directory and os.path.isdir(release_directory):
+            bak_dir = os.path.join(release_directory, 'bak')
             if not os.path.isdir(bak_dir):
                 os.mkdir(bak_dir)
             release_bak_dir = os.path.join(bak_dir, str(date.today()))